Migrate getPreviousQuestions in message class
